+++ src/main/java/com/koreplan/data/repository/DataRepository.java
@@ -14,7 +14,7 @@ public interface DataRepository extends JpaRepository<DataEntity,Long> {
 	//-----------------카테고리별 데이터 탐색------------//
 	List<DataEntity> findByC1Code(String C1Code);
 	List<DataEntity> findByC2Code(String C2Code);
-	List<DataEntity> findByC3Code(String C2Code);
+	List<DataEntity> findByC3Code(String C3Code);
 	//-----------------------------------------------
 	
 	//------------------지역별 데이터 탐색 -----------------//
