+++ src/main/java/com/koreplan/dto/search/ErrorResponseDto.java
@@ -1,4 +1,4 @@
-package com.koreplan.data.dto;
+package com.koreplan.dto.search;
 
 import java.time.LocalDateTime;
