+++ src/main/java/com/koreplan/entity/a/aa.java
@@ -1,0 +1,5 @@
+package com.koreplan.entity.a;
+
+public class aa {
+
+}
