+++ src/main/java/com/koreplan/service/search/FFfhdsjkah.java
@@ -1,0 +1,5 @@
+package com.koreplan.service.search;
+
+public class FFfhdsjkah {
+
+}
